+++ sierra_adapter/sierra_reader/src/test/scala/uk/ac/wellcome/platform/sierra_reader/flow/SierraRecordWrapperFlowTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_reader.sink
+package uk.ac.wellcome.platform.sierra_reader.flow
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
